LayoutTests/accessibility/scroll-to-make-visible-iframe-expected.txt

Tests that scrolling to make an element visible successfully scrolls an iframe.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



PASS frameWindow.pageYOffset is 0
PASS frameWindow.pageYOffset >= minYOffset is true
PASS frameWindow.pageYOffset <= maxYOffset is true
PASS frameWindow.pageYOffset >= minYOffset is true
PASS frameWindow.pageYOffset <= maxYOffset is true
PASS successfullyParsed is true

TEST COMPLETE
